refactor: add authentication middleware to routes for listing management and user actions

+++ routes/web.php
@@ -11,31 +11,31 @@ Route::get('/', [ListingController::class, 'index']);
 Route::get('/listings/{listing}', [ListingController::class, 'show'])->where('listing', '[0-9]+');
 
 // Show create listing form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store listing
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // Show register/create form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->name('register')->middleware('guest');
 
 // Create new user
 Route::post('/users', [UserController::class, 'store']);
 
 // Show login form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Login user
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
 
 // Logout
-Route::post('/logout', [UserController::class, 'destroy']);
+Route::post('/logout', [UserController::class, 'destroy'])->middleware('auth');
